+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "caniemail",
-  "version": "0.1.0-beta3",
+  "version": "0.1.0-beta4",
   "description": "HTML and CSS Feature Support for Email Clients from caniemail.com",
   "license": "MIT",
   "repository": "avigoldman/caniemail",
